pyproject.toml
[project]
name = "nuner-support-bot"
version = "1.0.0"
description = "Telegram Support Ticket Bot"
requires-python = ">=3.11"
dependencies = [
    "flask>=3.1.2",
    "pytelegrambotapi>=4.28.0",
    "python-dotenv>=1.1.1",
    "requests>=2.31.0"
]